refactor: components atoms to flat 변경 및 수정
