+++ src/components/MajorCompanys/CompanyItem.tsx
