+++ src/components/Sidebar/index.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/react'
-import MainLogo from 'components/atoms/MainLogo'
-import SidebarItem from 'components/atoms/SidebarItem'
-import Text from 'components/atoms/Text'
+import MainLogo from 'components/common/MainLogo'
+import SidebarItem from 'components/common/SidebarItem'
+import Text from 'components/common/Text'
 import { SectionItem, SectionList } from 'hooks/api/useGetSections'
 import media from 'lib/styles/media'
 import palette from 'lib/styles/palette'
