+++ src/components/common/AnimationItem.tsx
