+++ src/components/common/Badge.tsx
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react'
-import Text from 'components/atoms/Text'
+import Text from 'components/common/Text'
 
 interface IBadge {
   color: string
