+++ src/components/common/Banner.tsx
