+++ src/components/common/ClapButton/index.tsx
