+++ src/components/common/ClapButton/templates/Medium/icons/ClapEmpty.tsx
