+++ src/components/common/ClapButton/templates/Medium/icons/ClapFull.tsx
