+++ src/components/common/ClapButton/templates/Medium/style.ts
