+++ src/components/common/MainLogo.tsx
