+++ src/components/templates/IntroSection.tsx
@@ -1,10 +1,10 @@
 import { css } from '@emotion/react'
 import { logger } from '@lubycon/logger'
-import Button from 'components/atoms/Button'
-import Text from 'components/atoms/Text'
-import StickyButton from 'components/molecules/StickyButton'
-import MajorCompanyList from 'components/organisms/MajorCompanyList'
-import Statistics from 'components/organisms/Statistics'
+import Button from 'components/common/Button'
+import StickyButton from 'components/common/StickyButton'
+import Text from 'components/common/Text'
+import MajorCompanyList from 'components/MajorCompanys'
+import Statistics from 'components/Statistics'
 import { Intro } from 'hooks/api/useGetSections'
 import { useRouterQuery } from 'hooks/useRouterQuery'
 import media, { mediaQuery } from 'lib/styles/media'
