+++ src/pages/course/[courseSlug]/[sectionSlug]/index.tsx
@@ -1,15 +1,15 @@
 import { css } from '@emotion/react'
 import { logger } from '@lubycon/logger'
 import { useWindowSize } from '@lubycon/react'
-import Badge from 'components/atoms/Badge'
-import Icon from 'components/atoms/Icon'
-import Text from 'components/atoms/Text'
-import ArticleCard from 'components/molecules/ArticleCard'
-import IconButton from 'components/molecules/IconButton'
-import SectionRow from 'components/molecules/SectionRow'
-import StickyButton from 'components/molecules/StickyButton'
-import MobileSectionHeader from 'components/organisms/MobileSectionHeader'
-import Sidebar from 'components/organisms/Sidebar'
+import ArticleCard from 'components/ArticleCard'
+import Badge from 'components/common/Badge'
+import Icon from 'components/common/Icon'
+import StickyButton from 'components/common/StickyButton'
+import Text from 'components/common/Text'
+import IconButton from 'components/IconButton'
+import MobileSectionHeader from 'components/MobileSectionHeader'
+import SectionRow from 'components/SectionRow'
+import Sidebar from 'components/Sidebar'
 import Layout from 'components/templates/Layout'
 import LayoutResponsive from 'components/templates/LayoutResponsive'
 import useSectionDetail, { Blog, Section } from 'hooks/api/useGetSectionDetail'
