+++ src/pages/course/[courseSlug]/index.tsx
@@ -1,7 +1,7 @@
 import { logger } from '@lubycon/logger'
 import { useWindowSize } from '@lubycon/react'
-import MobileSectionHeader from 'components/organisms/MobileSectionHeader'
-import Sidebar from 'components/organisms/Sidebar'
+import MobileSectionHeader from 'components/MobileSectionHeader'
+import Sidebar from 'components/Sidebar'
 import IntroSection from 'components/templates/IntroSection'
 import Layout from 'components/templates/Layout'
 import LayoutResponsive from 'components/templates/LayoutResponsive'
