+++ src/pages/index.tsx
@@ -1,8 +1,8 @@
 import { logger } from '@lubycon/logger'
-import Banner from 'components/atoms/Banner'
-import Header from 'components/atoms/Header'
-import CourseList from 'components/organisms/CourseList'
-import MainFooter from 'components/organisms/MainFooter'
+import Banner from 'components/common/Banner'
+import Header from 'components/common/Header'
+import CourseList from 'components/CourseCard/CourseList'
+import MainFooter from 'components/Footer'
 import Layout from 'components/templates/Layout'
 import LayoutResponsive from 'components/templates/LayoutResponsive'
 import useGetCoruse from 'hooks/api/useGetCoruse'
